.docker/media_server/manager/requirements.txt
flask
requests
pymongo
ffmpeg-python